Merge PJT5 Branch into Master Branch

See merge request nts-2020-2nd-team5/jeon_kyung_yoon!25

+++ reservation/src/main/java/kr/or/connect/reservation/dao/DisplayInfoImageDao.java
@@ -23,7 +23,7 @@ public class DisplayInfoImageDao {
 		this.jdbc = new NamedParameterJdbcTemplate(dataSource);
 	}
 
-	public DisplayInfoImage selectUsingDisplayInfoId(Integer displayInfoId) {
+	public DisplayInfoImage selectUsingDisplayInfoId(Long displayInfoId) {
 		Map<String, ?> params = Collections.singletonMap("displayInfoId", displayInfoId);
 		return jdbc.queryForObject(SELECT_USING_DISPLAY_INFO_ID, params, rowMapper);
 	}
